Updates involving the search text field.
-Commented in gui/internal/edit/ClassifiableFilterEdit to document that
we should investigate how that class works more. It seems to populate
the pop-up that the filter button makes; we will want to point to an
action in that pop-up, which is without a doubt controlled by this
class.

-Documentation to gui/internal/edit/fields/SearchTextField which is the
JTextField we are using in our enhancement. We might need to adjust the
length of it in terms of DEFAULT_LENGTH and not a constant value.

-Added the SearchTextField to the ResourceSelection menu; it's non
functional, but you can type in it! :)