Moderate progress made!
I made a whole lot of changes. Enough that this might be better in the
experimental branch, but I pulled from master so I don't know how to
push to experimental. Everything still runs normally. The code
generating the popup on search is currently commented out, but can
easily be re-enabled.

The resource selection filter button is back, for easier testing and
debugging. It can be quickly removed as well. I just wanted to make sure
this is saved somewhere. I think I'm on to something...