Now, filter popups after you do a search.
For resources that have multiple attributes (person: first name,
surname, email...) it performs an AND search; the popup will allow the
user to change that, so they can eliminate one.

For example: User searches John (probably a first name)
The popup will come up, and they can discard the rules for last name and
email.

At the moment, we are now going to figure out how to make the AND search
an OR search so that the search is even more optimal.

SUCCESS!