Added reset feature for when you try to make a new search.
When you make a new search, it clears out the last search and then puts
in the new one.

Seems pretty obvious for a search feature but it didn't automate that
early. Now it's all set.

Tomorrow's goal: Try to get the AND search between different attributes
to an OR search.